+++ lib/screen/home.dart
@@ -1,0 +1,96 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import './product.dart';
+import './topic2.dart';
+import '../page.dart';
+
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Container(
+        padding: EdgeInsets.all(32.0),
+        child: Center(
+          child: ListView(
+            children: <Widget>[
+              GestureDetector(
+                onTap: (){
+                  var mainPageState = context.findAncestorStateOfType<Pages>();
+                  mainPageState?.onItemTapped(2);
+                },
+                child: Container(
+                    margin: EdgeInsets.all(10), width: 350, height:200 ,
+                    decoration: BoxDecoration(
+                        image: const DecorationImage(
+                          image: AssetImage('assets/images/coffee.jpg'),
+                          fit: BoxFit.cover,
+                        )
+                    ),
+                    child: Text(
+                        'order',
+                        style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: 24,
+                            color: Colors.white
+                        )
+                    ),
+                    alignment: Alignment.center
+                ),
+              ),
+              GestureDetector(
+                onTap: (){
+                  Navigator.push(
+                      context, MaterialPageRoute(builder: (context) => ProductPage()));
+                },
+                child: Container(
+                    margin: EdgeInsets.all(10), width: 350, height:200 ,
+                    decoration: BoxDecoration(
+                        image: const DecorationImage(
+                          image: AssetImage('assets/images/IMG_8836.jpg'),
+                          fit: BoxFit.cover,
+                        )
+                    ),
+                    child: Text(
+                        'Product',
+                        style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: 24,
+                            color: Colors.white
+                        )
+                    ),
+                    alignment: Alignment.center
+                ),
+              ),
+              GestureDetector(
+                onTap: () {
+                  Navigator.push(
+                      context, MaterialPageRoute(builder: (context) => Topic2Page()));
+                },
+                child: Container(
+                    margin: EdgeInsets.all(10), width: 350, height:200 ,
+                    decoration: BoxDecoration(
+                        image: const DecorationImage(
+                          image: AssetImage('assets/images/IMG_8837.jpg'),
+                          fit: BoxFit.cover,
+                        )
+                    ),
+                    child: Text(
+                        'Topic2',
+                        style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: 24,
+                            color: Colors.white
+                        )
+                    ),
+                    alignment: Alignment.center
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
