+++ lib/screen/inbox.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class InboxPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Inbox'),
+      ),
+      body: Container(
+        padding: EdgeInsets.all(32.0),
+        child: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Text('To be continued...'),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
