+++ lib/screen/product.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class ProductPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Product'),
+      ),
+      body: Container(
+        padding: EdgeInsets.all(32.0),
+        child: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Text('To be continued...'),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
